test: move parser exception logging test into test_extractor.py
